deribit: convert to TS
