+++ src/CandlePeriod.ts
@@ -3,10 +3,12 @@ export enum CandlePeriod {
     _2m = "_2m",
     _3m = "_3m",
     _5m = "_5m",
+    _10m = "_10m",
     _15m = "_15m",
     _30m = "_30m",
     _1h = "_1h",
     _2h = "_2h",
+    _3h = "_3h",
     _4h = "_4h",
     _6h = "_6h",
     _8h = "_8h",
